# os/LevelDBStore: faster LevelDBTransactionImpl::set